Specs pass without ParseTree so why depend on it
The problem with depending on ParseTree is that
it's not supported on all platforms.

dm-sweatshop for example depends on randexp and
since randexp depended on ParseTree there was
no way to override the platforms to use ParseTree
with, in dm-sweatshop's Gemfile. dm-sweatshop
indeed requires ParseTree to provide specific
functionality. Depending on ParseTree explicitly
in dm-sweatshop's Gemfile, means that we can limit
the platforms ParseTree will be installed on by
using bundler's :platform => :mri_18 option. Code
in dm-sweatshop can then rescue from a LoadError
from ParseTree, and will disable some features,
but will still be usable.